requirements.txt
langchain
openai
requests
openrouter
python-dotenv
langchain_openai
fpdf